requirements.txt
sentence-transformers
numpy
faiss-cpu
python-dotenv
fastapi
uvicorn
